Tests cleanup
